SaveSystemCommit

+++ Assets/ScriptableObjects/PlayerData.cs
@@ -3,7 +3,7 @@ using UnityEngine;
 [CreateAssetMenu(fileName = "PlayerData", menuName = "Maze/PlayerData")]
 public class PlayerData : ScriptableObject
 {
-    public int health;
-    public int damage;
-    public float moveSpeed;
+    public  float health;
+    public  float damage;
+    public  float moveSpeed;
 }
